Disable preemptive DB check
Sequel connections have changed to immediately connect to the DB when created.
As an early warning mechanism.

This causes problems for our DB create and drop rake tasks. As we load the
environment/app before running these tasks. It'll connect before we issue the
command:

 * on `create` the app connection will blow up as the database doesn't yet exist
 * on `drop` the drop rake task will blow up because the app is holding a
   connection to the database.

We can disable this new preemptive DB check by passing `test: false` when
creating the connection.

/ht @orien